Move Python modules into examples/python
