+++ examples/python/stream_module.py
