+++ examples/python/streams_log.py
